openai_moderation/requirements.txt
psycopg2-binary
python-dotenv
clean-text
openai